backend/requirements.txt
flask
flask-cors
groq
requests
python-dotenv